project/requirements.txt
customtkinter==5.2.0
pillow==10.0.0
fpdf2==2.7.8